refactor(post): migration posts from Mongo to PostgreSQL

+++ src/modules/blogging.platform/application/post.service.ts
@@ -1,44 +1,52 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
-import { InjectModel } from '@nestjs/mongoose';
+import { Injectable } from '@nestjs/common';
 import { PostRepository } from '../infrastucture/post.repository';
-import { Post, PostDocument, PostModelType } from '../domain/post.entity';
+import { Post } from '../domain/post.entity';
 import { PostInputDto } from '../dto/input/post.input.dto';
 import { BlogQueryRepository } from '../infrastucture/query/blog.query.repository';
+import { PostByBlogInputDto } from '@modules/blogging.platform/dto/input/post.by.blog.input.dto';
 import { DomainException } from '@core/exceptions/domain.exception';
 import { DomainExceptionCode } from '@core/exceptions/domain.exception.code';
+import { PostParamsIdInputDto } from '@core/dto/input/post.params.id.input.dto';
 
 @Injectable()
 export class PostService {
     constructor(
         private postRepository: PostRepository,
         private blogQueryRepository: BlogQueryRepository,
-        @InjectModel(Post.name) private PostModel: PostModelType,
     ) {}
 
     async create(inputItem: PostInputDto): Promise<string> {
         const blogName
             = (await this.blogQueryRepository.findByIdWithCheck(inputItem.blogId)).name
-        const newPost: PostDocument = this.PostModel.createInstance(inputItem, blogName);
-        await this.postRepository.save(newPost);
-        return newPost._id.toString();
+        const newPost: Post = Post.createInstance(inputItem, blogName);
+        await this.postRepository.savePost(newPost);
+        return newPost.id.toString();
     }
 
-    async edit(id: string, editData: PostInputDto): Promise<void> {
+    async edit(idDto: PostParamsIdInputDto, editData: PostByBlogInputDto): Promise<void> {
 
-        const post: PostDocument = await this.postRepository.findById(id);
+        const post: Post | null = await this.postRepository.findByIdWithoutBlog(idDto.id);
 
-        const blogName: string
-            = (await this.blogQueryRepository.findByIdWithCheck(editData.blogId)).name
-        post.update(editData, blogName);
-        this.postRepository.save(post);
+        if(!post || post.blogId !== +idDto.blogId)
+            throw new DomainException({
+                message: 'post with ${id}  not found',
+                code: DomainExceptionCode.NotFound});
+
+        post.update(editData);
+        this.postRepository.savePost(post);
         return;
     }
 
-    async delete(id: string): Promise<void> {
-        const post: PostDocument = await this.postRepository.findById(id);
+    async delete(idDto: PostParamsIdInputDto): Promise<void> {
+        const post: Post | null = await this.postRepository.findByIdWithoutBlog(idDto.id);
+
+        if(!post || post.blogId !== +idDto.blogId)
+            throw new DomainException({
+                message: 'post with ${id}  not found',
+                code: DomainExceptionCode.NotFound});
 
         post.delete();
-        this.postRepository.save(post);
+        this.postRepository.savePost(post);
         return;
 
     }
